+++ nodes/Medusa/shared/router.ts
@@ -6,11 +6,15 @@ import {
 } from 'n8n-workflow';
 import { productOperations } from '../resources/product';
 import { variantOperations } from '../resources/productVariant';
+import { categoryOperations } from '../resources/productCategory';
+import { collectionOperations } from '../resources/productCollection';
 import type { OperationHandlers } from './types';
 
 const resources: Record<string, OperationHandlers> = {
 	product: productOperations,
 	productVariant: variantOperations,
+	productCategory: categoryOperations,
+	productCollection: collectionOperations,
 };
 
 /**